Add Webpack config to main process
- Add new npm scripts
- Extract functions in separate libs (menu, etc.)
- Clear ipc channels before reactivating window on macos as mainWindow is not overwritten
- add custom hot reload code
- use webpack to scrap dev and testing code from prod bundle
