+++ package.json
@@ -23,10 +23,10 @@
     "test": "xvfb-maybe ts-mocha --paths -p test/tsconfig.spec.json --timeout 20000 test/suites/**/*.spec.ts",
     "clean-dist": "rm -rf ./dist",
     "clean-packages": "rm -rf ./packages",
-    "build:watch:main": "tsc -w -p ./src/main/tsconfig.json",
-    "build:dev:main": "tsc -p ./src/main/tsconfig.json",
-    "build:ci:main": "tsc -p ./src/main/tsconfig.json --sourceMap false",
-    "build:prod:main": "tsc -p ./src/main/tsconfig.json --sourceMap false",
+    "build:watch:main": "webpack --config ./src/main/webpack.config.js --mode development --watch",
+    "build:dev:main": "webpack --config ./src/main/webpack.config.js --mode development",
+    "build:ci:main": "webpack --config ./src/main/webpack.config.js --mode production --env isTesting=1",
+    "build:prod:main": "webpack --config ./src/main/webpack.config.js --mode production",
     "build:watch:renderer": "ng build --configuration=development --watch --output-hashing=all",
     "build:dev:renderer": "ng build --configuration=development --output-hashing=all",
     "build:ci:renderer": "ng build --prod --configuration=ci",
@@ -35,7 +35,7 @@
     "build:dev:all": "npm run clean-dist && npm run build:dev:renderer && npm run build:dev:main",
     "build:ci:all": "npm run clean-dist && npm run build:ci:renderer && npm run build:ci:main",
     "build:prod:all": "npm run clean-dist && npm run build:prod:renderer && npm run build:prod:main",
-    "start": "electron ./dist --serve --remote-debugging-port=8888",
+    "start": "electron ./dist --remote-debugging-port=8888",
     "package:win": "npm run clean-packages && electron-builder --publish never --win --config electron-builder.json",
     "package:linux": "npm run clean-packages && electron-builder --publish never --linux --config electron-builder.json",
     "package:mac": "npm run clean-packages && electron-builder --publish never --mac --config electron-builder.json"
@@ -51,7 +51,6 @@
     "@mockoon/commons": "2.2.0",
     "@mockoon/commons-server": "2.2.1",
     "axios": "0.21.1",
-    "electron-is-dev": "2.0.0",
     "electron-json-storage": "4.5.0",
     "electron-log": "4.3.5",
     "electron-window-state": "5.0.3",
@@ -61,7 +60,7 @@
     "semver": "7.3.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.12",
+    "@angular-devkit/build-angular": "0.1102.13",
     "@angular-eslint/builder": "4.2.0",
     "@angular-eslint/eslint-plugin": "4.2.0",
     "@angular-eslint/eslint-plugin-template": "4.2.0",
@@ -100,13 +99,13 @@
     "chai": "4.3.4",
     "chai-as-promised": "7.1.1",
     "chai-exclude": "2.0.3",
+    "chokidar": "3.5.1",
     "concurrently": "6.1.0",
     "date-fns": "2.21.3",
     "dotenv": "9.0.1",
     "electron": "12.0.7",
     "electron-builder": "22.11.1",
     "electron-notarize": "1.0.0",
-    "electron-reloader": "1.2.1",
     "eslint": "7.26.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-chai-friendly": "0.7.1",
@@ -126,12 +125,16 @@
     "rxjs": "6.6.7",
     "spectron": "14.0.0",
     "timers": "0.1.1",
+    "ts-loader": "8.3.0",
     "ts-mocha": "8.0.0",
     "tsconfig-paths": "3.9.0",
     "tslib": "2.2.0",
     "typescript": "4.1.5",
     "uuid": "8.3.2",
     "wait-on": "5.3.0",
+    "webpack": "4.44.2",
+    "webpack-cli": "4.7.0",
+    "webpack-node-externals": "3.0.0",
     "xvfb-maybe": "0.2.1",
     "zone.js": "0.11.4"
   }
