+++ src/main/constants/ipc.constants.ts
@@ -1,0 +1,27 @@
+export const IPCListenerChannels = [
+  'APP_APPLY_UPDATE',
+  'APP_DISABLE_EXPORT',
+  'APP_ENABLE_EXPORT',
+  'APP_LOGS',
+  'APP_OPEN_EXTERNAL_LINK',
+  'APP_QUIT',
+  'APP_SET_FAKER_OPTIONS',
+  'APP_UPDATE_ENVIRONMENT',
+  'APP_WRITE_CLIPBOARD'
+];
+
+export const IPCHandlerChannels = [
+  'APP_GET_MIME_TYPE',
+  'APP_GET_PLATFORM',
+  'APP_OPENAPI_DEREFERENCE',
+  'APP_OPENAPI_VALIDATE',
+  'APP_READ_CLIPBOARD',
+  'APP_READ_FILE',
+  'APP_READ_JSON_DATA',
+  'APP_SHOW_OPEN_DIALOG',
+  'APP_SHOW_SAVE_DIALOG',
+  'APP_START_SERVER',
+  'APP_STOP_SERVER',
+  'APP_WRITE_FILE',
+  'APP_WRITE_JSON_DATA'
+];
