+++ src/main/libs/hot-reload.ts
@@ -1,0 +1,25 @@
+import { watch } from 'chokidar';
+import { app, BrowserWindow } from 'electron';
+import { basename } from 'path';
+
+export const hotReload = (mainWindow: BrowserWindow) => {
+  let relaunching = false;
+
+  const watcher = watch(__dirname, { ignored: '**/*.map' });
+
+  app.on('quit', () => {
+    watcher.close();
+  });
+
+  watcher.on('change', (path) => {
+    if (['app.js', 'preload.js'].includes(basename(path))) {
+      if (!relaunching) {
+        app.relaunch();
+        app.exit(0);
+        relaunching = true;
+      }
+    } else {
+      mainWindow.webContents.reloadIgnoringCache();
+    }
+  });
+};
